+++ src/main/java/congregamystica/integrations/congregamystica/blocks/inventories/gui/GuiPatternProvider.java
@@ -6,7 +6,6 @@ import congregamystica.integrations.congregamystica.blocks.tiles.TileArcanePatte
 import net.minecraft.client.gui.inventory.GuiContainer;
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.inventory.Container;
 import net.minecraft.inventory.Slot;
 import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.fml.relauncher.Side;